+++ package.json
@@ -8,8 +8,8 @@
         "start": "next start",
         "lint": "next lint",
         "format": "prettier --write .",
-        "test": "jest --watch",
-        "test:ci": "jest --ci",
+        "test": "jest --ci",
+        "test:watch": "jest --watch",
         "postinstall": "prisma generate"
     },
     "dependencies": {
