+++ src/pages/my-account/index.tsx
@@ -3,7 +3,7 @@ import toast from 'react-hot-toast';
 import {
     useGetProfileQuery,
     useDeleteAccountMutation
-} from '@/features/manage/manage.api';
+} from '@/features/manage/manageApi';
 import { Meta } from '@/components/Meta/Meta';
 import { Container } from '@/components/Container/Container';
 import { Title } from '@/components/Title/Title';
@@ -14,7 +14,7 @@ import { AccountSettingsCard } from '@/features/manage/AccountSettingsCard/Accou
 import { GetServerSideProps } from 'next';
 import { getServerSession } from 'next-auth';
 import { getRunningQueriesThunk } from '@/redux/api';
-import { getProfile } from '@/features/manage/manage.api';
+import { getProfile } from '@/features/manage/manageApi';
 import { wrapper } from '@/redux/store';
 import { authOptions } from '@/pages/api/auth/[...nextauth]';
