+++ src/pages/reset-password/[token].tsx
@@ -1,6 +1,6 @@
 import { useRouter } from 'next/router';
 import toast from 'react-hot-toast';
-import { useResetPasswordMutation } from '@/features/account/account.api';
+import { useResetPasswordMutation } from '@/features/account/accountApi';
 import { Meta } from '@/components/Meta/Meta';
 import { Container } from '@/components/Container/Container';
 import { Title } from '@/components/Title/Title';
