+++ src/pages/update-profile/index.tsx
@@ -3,7 +3,7 @@ import toast from 'react-hot-toast';
 import {
     useGetProfileQuery,
     useUpdateProfileMutation
-} from '@/features/manage/manage.api';
+} from '@/features/manage/manageApi';
 import { Meta } from '@/components/Meta/Meta';
 import { Container } from '@/components/Container/Container';
 import { Title } from '@/components/Title/Title';
@@ -16,7 +16,7 @@ import {
 import { GetServerSideProps } from 'next';
 import { getServerSession } from 'next-auth';
 import { getRunningQueriesThunk } from '@/redux/api';
-import { getProfile } from '@/features/manage/manage.api';
+import { getProfile } from '@/features/manage/manageApi';
 import { wrapper } from '@/redux/store';
 import { authOptions } from '@/pages/api/auth/[...nextauth]';
